+++ Src/Library/Endpoint/Endpoint.Base.cs
@@ -50,7 +50,7 @@ public abstract class BaseEndpoint : IEndpoint
     [NotImplemented]
     public virtual void Configure() => throw new NotImplementedException();
 
-    public virtual void Verbs(params Http[] methods) => throw new NotImplementedException();
+    public virtual void Verbs(params string[] methods) => throw new NotImplementedException();
 
     protected virtual void Group<TEndpointGroup>() where TEndpointGroup : notnull, Group, new() => throw new NotImplementedException();